+++ apps/api/src/api/user/commands/update-trial-status.command.ts
@@ -1,0 +1,6 @@
+import { UserModel } from "../models/user.model"
+
+export async function updateTrialStatus(userId: string) {
+  await UserModel.findByIdAndUpdate(userId, { isTrialAvailable: false })
+  return true
+}